Moved mappers and enums to seperate files
